moved all Cook model tests to accounts
